Remove LITERAL and NEXT instructions, and A register
This is a major change to the execution mechanism. SMite is now a pure
byte-code machine, and literal numbers have a new encoding scheme that uses
much of the opcode space.

Remove the FROM command from the shell, which is no longer needed.

Also do some tidy-up, e.g. find remaining instances of “EP” in documentation
and rename to “PC”.

Now we no longer have NEXT, add NOP.