Drop support for object files of opposite ENDISM to host
With the change to byte-code, the old scheme no longer worked: reversing
words on load would put the bytes in the wrong order, since I is loaded a
byte at  time, not a word at a time.

Add a note to TODO on how to support opposite-ENDISM files if desired.