Overhaul API to make it FFI friendly
This is a prelude to rewriting the shell in Python, but similar
considerations apply to use with most scripting languages.

Mostly, we turn macros into functions or constants, so that we can bind them
more easily (e.g. in Python via ctypes), rather than having to use a binding
generator that parses header files, such as SWIG.

Also add most of debug.c to contrib/python-setup.py (currently compiles, but
mostly untested).