Duplicate stack APIs for return and data stacks
Thus only need to pass a state argument in rather than registers and
pseudo-registers, thus making the APIs easier to bind via FFIs, in
particular because it is not necessary to pass SMite-internal variables in
as by-reference return values.

This API duplication will go away when we get rid of the data stack later.