Travis: use xenial explicitly, and fix use of latest GCC (8)
The compiler was not being set up properly.

Add settings to use ASAN, but comment them out. The current recent GCCs in
Travis have broken support for ASAN (option passing to the linker fails),
and the workaround of using the gold linker also fails, because libtool
doesn’t support that. Le sigh.