Streamline instruction set
Remove CALL_NATIVE: it is not portable between implementations, so there’s
no point having them in the core definition. Also, I’ve never used it. Hence
remove PUSH_NATIVE_POINTER_SIZE.

Remove PUSH_PC, PUSH_MEMORY and PUSH_STACK_SIZE. The first is unnecessary,
and the other two are properties of the system, not the “CPU”.

Add ARSHIFT and ROTATE.

Rename INVERT to NOT, PUSH_* to GET_* and STORE_* to SET_*.

Change encoding slightly so that action opcodes have top two bits clear (so
that byte 0x00 is HALT, and hence runaway code tends to stop fast).

Reorder the opcodes to create an aesthetically more pleasing table, and so
that HALT has opcode 0 for fast-fail on bad code.