src/aux.h: fix UNCHECKED_LOAD/STORE_STACK
These were (undetected) accessing the wrong stack locations, starting at
STACK_DEPTH and working upwards (i.e. overflowing) rather than down. Fix
this.

Add a FIXME to aux.h to make the macros depend on ENDISM, so that native
types are stored on the stack in native order. (This just seems the nice
thing to do; there’s no specific need for it, though it might be possible to
implement the loads and stores more efficiently given this constraint.) The
current code is correct for ENDISM=0.