smite_core/action_gen.py: fix spec & use of COUNT
COUNT is actually one less than the number of unnamed items on the stack.
This may seem odd, but means that the case for 0 arguments to stack
operations is defined and consistent, and, for SWAP and ROTATE_UP/DOWN, a
no-op. The only obvious alternatives are either to make zero an error, or to
make it the same as 1. Neither seems terribly satisfactory. Forth uses the
same 0-based numbering (as a vote in our favour!).