Mark `S` as "restrict" in `mit_single_step()` and `mit_run()`. Also in `mit_run_specialized` and `mit_run_trace`.
I decided it was pointless to do the same in the static inline functions in "mit.h".
If they are inlined it is unnecessary, and if they are not it is futile.

mit.h: Mark `memory` and `stack` as "restrict" in "struct mit_state".
storage.c: smite_realloc: Add "restrict" to fix warnings.