src/main.c: generate with gen-main
All features are in src/feature/*.py

The feature files start with a two-digit number so they can be ordered,
and are read in lexical order (according to Python 3’s “sorted”).

As a side-effect of doing the code generation in Python rather than cpp, we
no longer need xvasprintf; for the one remaining use, use sprintf and a
conservatively-sized static buffer.