smite_core/instruction_gen.py: update S->STACK_DEPTH only after action routine
Suggested by @apt1002, to make it easier for the compiler to optimise the
update. Also removes the need to cache the original value in case of error.

In ext.py, the Library class must manually pop the function code, because
STACK_DEPTH being updated only after an action routine means that actions
routines can no longer be nested.