Simplify specification and implementation of instruction fetch
Thanks @apt1002, for noting a nit that led me to reconsider.

Move all the instruction fetch semantics to the main loop in the spec, where
there is now an if that either shifts I right or loads it from PC.

As a result, NEXT becomes NOP, and BRANCH, BRANCHZ and CALL no longer
perform instruction fetch. This does not affect the value put in BAD when an
error is raised.

This also simplifies the assembler and the branch test.