Install registers.h, and make installed headers work
Previously, smite.h contained #include "registers.h". This would
fail, as registers.h was not installed. But even after installing
registers.h, it would not have worked, because the installed path
is "smite/registers.h". Hence, put all the header files under
a subdirectory "smite", but because that is the name of the front-
end binary, put that itself under a subdirectory "include".