Separate mit_core from Python bindings and move type-sizes into features
Move the type-sizes machinery into features: it’s only needed for extra
instructions.

Having done this, the most straightforward way to make the Python bindings,
which need the extra instructions’ opcodes, to work again was to break the
link between mit_core and the Python bindings (in the form of modules
symlinked from the former into the latter), and instead generate a Python
opcodes module, mirroring the C opcodes headers.