Merge gen-trace and gen-predictor (fix #42)
As a result, instructions_predictor.json must be checked in, both to avoid a
circular dependency with libmitfeatures.la, and because it can only be
produced by a mit that can run pForth, at present.

Also, gen-instructions has to generate frequency information from the
predictor, as there is no longer an instruction trace.

Adapt mit_core.opcode_frequency to read a predictor. Abandon generation of
frequencies for opcode pairs.