Fix missing type size information in gen-extra-instructions
The problem was that registers are of type mit_uword, but since the register
accessor functions are auto-generated without stack effects (because some of
them alter the stack, so the auto-generated code for stack access would be
wrong), the type `mit_uword` does not appear in the list of types to find
the size for.

Hence, add `mit_uword` to the hard-wired list (we know it is of size 1), and
have gen-extra-instructions update this list from `type_sizes` rather than
overwrite it.