Make type-sizes machinery report errors properly (fix #228)
The fix is not as originally envisaged: the job done by autoconf’s
AC_CHECK_TYPE (called by AC_CHECK_SIZEOF) is complex, and I don’t want to
replicate the code. Hence, I want to continue to use autoconf, and hence,
it’s important for performance reasons to check all the type sizes in one
go.

The important thing is that missing types must cause an error rather
than (as before, despite comments to the contrary) silently compiling
incorrect code.

To fix this, while allowing type sizes to be unknown when they are being
generated by gen-type-sizes, or unneeded, as in gen-extra-opcodes, add a
special value instruction_gen, TYPE_SIZE_UNKNOWN. This defaults to
None (which will cause an error), but can be set to 0 (which won’t).

Making this change uncovered a missing type: LibInstruction uses
“mit_word *” in the VM register accessor functions, which (necessarily) have
None as their StackPictures, so gen-type-sizes does not find this type.
Hence, add a field “extra_types” to Library, which allows a library to
specify required types (of stack items) that are not specified in
StackPictures.

This uncovered another bug, to do with register pointer types that are
marked “restrict”; remove the return_type field of Register, and instead add
a utility function code_util.unrestrict that removes “restrict” (correctly!)
where needed. (Before, it wasn’t always removed, and when it was, trailing
spaces remained, so the type size could still not be correctly looked up.)