Python: tweak some imports
To my surprise, “import os.path” seems to import all of “os” (as
load_object.py imported os.path, but referenced other parts of os). Hence,
correct the import in load_object.py to “import os”, since we really do want
other bits of os, and in mit-dump.in change to “import os”, since that’s
what we seem to be getting anyway.