Overhaul optional tests: add “make check-bench”
Make it possible to run optional tests without reconfiguring. Currently, all
the optional tests are benchmarks, so move them to a separate category,
“BENCH_TESTS”, with a convenience target “make check-bench”.

Refactor the code to run tests of pForth building itself. A single build is
run as part of the regular tests, and a tenfold test with optimized and
non-optimized interpreter for “make check-bench”.