Python: simplify and fix State.step()
State.step() was not stopping when PC equalled the addr argument if that
occurred when auto_NEXT=True and a new instruction word was fetched at that
point.

Simplify the code for auto_NEXT to simply decrement the instruction count
“done”, as the reason for auto_NEXT is to avoid counting implicit “NEXT”
instructions.

This removes duplication and incidentally fixes the above bug.

Also output PC in hex when trace=True.