Add step N times/until pc=addr as C feature (issue #244)
This feature is used by Python step/trace (the latter is a new function,
instead of the trace flag to step(); tracing is still slow, but that’s OK as
typically it will be I/O bound anyway).

This makes it much faster to step to a breakpoint or for millions of
instructions.