Change C and Python APIs to specify memory size of VM in bytes (fix #252)
(Not words, which is confusing.)

Stack sizes are still in words, however, since we’re really specifying
a number of stack items. Make this clear by naming the various parameters
and pseudo-registers accordingly.

mit_realloc_memory() return an error if the memory size requested is
unaligned.

Take advantage of this change to simplify some test code.