features/predictor.c: allocate memory for predictor dynamically
This started as a simple fix of a FIXME: rather than hard-wiring the number
of opcodes (which is not available conveniently), use the upper-limit of
2^8. This increased the amount of memory statically allocated by the
predictor (only) eight-fold, which in turn broke linking of libmitfeatures,
as some relocations no longer worked.

This drew my attention to a large amount of memory being allocated
statically; so instead, I changed the API to allocate it dynamically (but
still globally).