simulate-jit: Introduce classes State, If and Goto and use them
State identifies a control-flow state. Not having to number these in
advance allows some separation of concerns.

Goto represents an unconditional jump to another State. Being able to
generate these allows some separation of concerns.

There is a new branch-tensioning pass that removes all the Gotos and
redundant Ifs.

Performance appears to be unchanged as it should be.