Use StackEffect directly in Instruction; add constant StackItems
Before, the arguments and results of an Instruction were stored as string
lists, and only passed to StackPicture/StackEffect in gen_case.

Although this simplified Instruction slightly, it also led to duplication,
for example computing the set of argument & result names in more than one
place, as well as using proxies such as the presence of `ITEMS` in the
arguments & results lists to diagnose a variadic instruction.

This refactoring allows the use of StackItem properties other than in
gen_case, and moves StackEffect and its components into their own module,
mit_core/stack.py.

Allow StackItems to be set to a constant. This ability will be used in the
next commit.