configure.ac: we actually need Python 3.6
simulate-jit uses random.choices. It would be inconvenient to fix, and given
how straightforward it is to install a newer Python where needed, seems
silly to keep policing ourselves in this way.

In future I may well require the latest stable major Python version, and
use a suitable distro-neutral method for installing it in Travis, to give a
documented recipe.