Remove the ‘bad’ register
It’s simply not that useful. Notably, pForth and the tests don’t use it.

The one feature that did rely on it was auto-extending stack and memory.
Auto-extending memory will not make sense when Mit is using the native
address space, and so it’s removed; auto-extending stack has been
simplified. What remains of the feature is folded back into gen-main.

Remove a stack check in gen-main that was never generated (and was
syntactically invalid).