Make extra (non-spec) registers writable
Specifically, the memory, memory_words, stack and stack_words members of
mit_state. This means they now have setter methods.

The motivation for this is to allow states to share a memory and/or
stack. (mit_new_state() creates a stackless or memoryless state when the
corresponding size passed is 0).