Make it easy to use a particular execution function in nested states
Add `run_fn` and `extra_state` members to `mit_state`. Make the extra
instruction `RUN` call `run_fn` rather than `mit_run`, allowing nested
states to naturally use the same run function. Add a type `mit_fn` for
this sort of function and document its use.

Remove the extra instructions for the specializer and profiler, as the
run_fn should typically be set outside the VM, and profiling, if desired,
could be re-implemented in Mit code; it’s not necessary to make the
specializer-specific mechanism available to VM code.