Fix implementation of HALT (fixes FIXME, adds another!)
instruction_gen pops instructions’ arguments after completing execution of
their code. This does not work for HALT, whose normal execution raises an
error, and therefore skips the popping of the arguments.

For now, deal with this as we do for the other extra instructions that
cannot be implemented automatically (those that deal with registers that
change during instruction execution). Add a FIXME to fix code generation for
this case.