Remaster specification in Python
Generate mit_spec.yaml from spec.py. This is preparatory to adding generated
`pushN` instructions, for which the source should be generating code, not
the written-out result.

Also add an argument to code_util.copyright_banner() to specify the comment
string, and use it to generate copyright banners for Python and YAML.