mit_core/params.py: move word_bit into spec.py
Note that MIT_WORD_BYTES still needs to be substituted for word_bytes, to
support configuring with a non-native architecture e.g. “gcc -m32” on x86.
This will cause the Python bindings to fail, but is the only way to avoid
using Docker on such systems, and `mit` will run fine.