Add immediate `jumpzi`, like `jumpi` but for `jumpz`.
In theory, `jumpz` could always be followed by other instructions, but I
think it is too ugly to make it different from `jump`.

I also considered replacing `jumpz` with `skipnz`, which zeroes `ir` if the
top-of-stack is true, but for common use cases this requires quite a few
instructions (in particular, `jumpz` is `skipnz 0 swap jumpz-with-address`,
which even with one-byte `push` will only allow a one-byte offset on 4-byte
Mit). Also, `skipnz` is awkward to implement efficiently, even on e.g. ARM
with condition codes (consider when `skipnz` is followed by a test such as
`ult`): lots of special cases required to do the obvious thing.

So, rather like `call` (which could be `get_pc 0 swap jump`, again ugly),
leave duplicated functionality as it is, because it matches CPU
architectures better.