src/Makefile.am: fix C dependencies
The C dependencies support does not work for .h files generated at build
time, as the dependencies files are generated while compiling C sources, so
the headers must already exist.

Rather than declare normal make-style dependencies of object files on source
files for this case, write source dependencies of C files on header files,
as in auto-generated dependencies files, and only add dependencies manually
on auto-generated files (before I was just trying to guess).

Thanks to @apt1002 for asking what was going on here and making me think
about it enough that I was able to work out the problem and, later, the
solution!