Rework the classes used to represent instructions (fix #354)
Add a new class Action, which represents the semantics of an instruction (or
similar sort of thing), without its encoding.

Use Action to define Instruction in spec.py, and ActionEnum to define all
the various instruction(-like) enumerations.

Because there is now an Instruction class to represent Mit instructions,
pluralize all the enumerations: Instructions, ExtraInstructions etc.