Specializer: improve `make [re]specialize`
Time the initial profile (it may be the best! particularly when
respecializing).

At the end of `repeat-specialize`, copy the best profile without preserving
its timestamp, and then add a final invocation of make to the `respecialize`
target, so that `mit` ends up rebuild with the fastest profile (that it
didn’t do this was confusing).