Make stack a stack of frames; modify `CALL` and add `RET`
This is a simpler version of the design on the `frame-stack` branch.

`mit_single_step()`, and the stack access APIs, and the extra instructions
to access the VM registers have been removed. There is a separate debugging
version of `mit_run_simple()`, `mit_run_break()`.

In Python, `State.step()` no longer supports `auto_NEXT` argument. It is no
longer possible (because no longer useful) to pickle a `State`. Moved
disassembler into its own module. Added a `next_pc()` method to `Assembler`
to get the effective `pc` after executing an instruction; used this in
branch.py.

pForth is not yet updated, so disable use of it and hence the
specializer (which in any case is obsolescent).