Simplify `mit_break()` design
Define `mit_callback_fn`, which is a function run in the context of a Mit
stack frame, and is passed all the necessary state. It can modify the stack.

Make `mit_break_fn` a `mit_callback_fn`.

Simplify `State.BreakHandler` to not try to guess which `next` instructions
to skip; leave that to the caller.

Call `mit_break_fn` right at the start of the main interpreter loop, not at
the end.

As a result, remove `Assembler.next_pc()`, as it’s no longer necessary; the
callback in `mit_test.py` now counts the number of significant instructions
executed itself, rather than relying on `handler.done`.