Make POP niladic (again)
The reason for giving it an argument was that it was trivial to pop `n`
items from the stack, so why not?

The reason for removing the argument is that it makes `pop` more symmetric
with `dup`: you can either add an item to the stack or remove one. It is
also more symmetric when mapped to registers: a new register is added to the
live set by `dup` and one is removed by `pop`.

Since by far the commonest use is to pop a single item, this should improve
code density.

Finally, now that the stack is divided into frames, it’s no longer useful to
discard many items at once (as it was to implement stack frames with a
continuous stack).