python/assembler.py: rework `jumprel()` to emit more compact code
Previously, if an immediate JUMP didn't fit in the current word, we
emitted a PUSHREL followed by an indirect JUMP. Now we skip to the next
word and emit an immediate JUMP. The old behaviour is used only for the
rare case when the offset doesn't fit in an immediate field of size
`WORD_BYTES-1`.

Also, rework tests/branch.py to test all assembly positions for jumprel,
and clarify the output a bit.