python/assembler.py: fix and improve `pushrel()`; add `pushrel_long()`
pushrel() would assemble the wrong offset if there was no room for it to
assemble an opcode in the current instruction word, as `instruction()` would
automatically move to the next word, but not, of course, recalculate the
offset.

Factor out `pushrel_long()` to replace the `force_long` flag to
`pushrel()`. This fixes an edge case.

Thanks to @apt1002 for pointing out an extra case, namely that for an offset
of 64 words it’s a good idea to skip to the next word anyway, as then it
will be possible to use a one-byte `pushreli_63` instruction, rather than
needing a whole word plus a byte for a `pushrel`.