mit.tex: improve presentation of instruction set (fixes #402)
Remove the table of types; they’re not useful, and are replaced by a few ad
hoc remarks about the interpretation of stack items (mostly where counts
must be treated as unsigned numbers).

Instead of type names with suffixes, use identifiers, as in `spec.py` (see
issue #405, to bring `spec.py` into better alignment with `mit.tex`).

Improve some instruction descriptions.

Add an extra column to the instruction encoding table, to clarify what goes
in the instruction bits. Use the instruction number, not the full opcode in
the instruction opcodes table, and add a new 1-entry table for the missing
opcode for `trap`.