Make “throw 0” an error (see issue #403)
`throw 0` was previously allowed, but the caller cannot distinguish it from
`ret`, so cannot tell whether to look for results underneath the status code
or not. Now things are more consistent: `throw` always throws an error, and
if the result of `catch` is 0, then there are results under the status code.

In the implementation, this removes the need for the error code to be
mangled when it was zero (which is a normal return from `setjmp()`, not an
error return).

It may well be desirable to implement non-local return in future, possibly
through a mechanism orthogonal to `catch`/`throw`.